upnote/config.py
from dotenv import load_dotenv
from os import environ

load_dotenv()


class Config(object):
    SECRET_KEY = environ.get('SECRET_KEY')

    SQLALCHEMY_DATABASE_URI = environ['DATABASE_URL']
    SQLALCHEMY_TRACK_MODIFICATIONS = False
